Isolated PostgreSQL test into PostgreSQL folder
